+++ src/adaptors.rs
@@ -3,17 +3,18 @@
 //! Bot adaptors are very similar to the [`Iterator`] adaptors: they are bots
 //! wrapping other bots to alter existing or add new functionality.
 //!
-//! E.g. [`AutoSend`] allows `await`ing requests directly, no need to use
-//! `.send()`.
-//!
 //! [`Requester`]: crate::requests::Requester
 
-/// [`AutoSend`] bot adaptor which allows sending a request without calling
-/// [`send`].
+/// [`AutoSend`] bot adaptor which used to allow sending a request without
+/// calling [`send`].
 ///
 /// [`AutoSend`]: auto_send::AutoSend
 /// [`send`]: crate::requests::Request::send
 #[cfg(feature = "auto_send")]
+#[deprecated(
+    since = "0.8.0",
+    note = "`AutoSend` is no longer required to `.await` requests and is now noop"
+)]
 pub mod auto_send;
 
 /// [`CacheMe`] bot adaptor which caches [`GetMe`] requests.
@@ -47,6 +48,7 @@ pub mod throttle;
 mod parse_mode;
 
 #[cfg(feature = "auto_send")]
+#[allow(deprecated)]
 pub use auto_send::AutoSend;
 #[cfg(feature = "cache_me")]
 pub use cache_me::CacheMe;